Modulo Veterinario y agenda terminado Junto con el fix de una migracion de la tabla cita, fix de el modulo vacuna

+++ src/main/java/com/Vet/VetBackend/agenda/web/dto/DetalleHorarioVeterinarioModificarReq.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class DetalleHorarioVeterinarioModificarReq {
     private int veterinarioId;
     private int diaId;
-    private Byte bloqueHorarioId;
+    private int bloqueHorarioId;
 }
